Breaks apart SheetSyncer into smaller modules
Why
---
SheetSyncer is complicateda and only downloads from the worksheet only.
It is not easy to reuse to upload to the worksheet.

This change addresses the need by
---------------------------------
Breaks apart SheetSyncer into modules that can be reused for uploading
and keeps all download logic in the Downloader class.
