+++ app/services/sheet_sync/uploader.rb
@@ -1,0 +1,4 @@
+module SheetSync
+  class Uploader
+  end
+end
