+++ lib/tasks/sync_reviews.rake
@@ -1,6 +1,6 @@
 namespace :reviews do
   desc "Syncs reviews with Google Sheet"
   task sync: :environment do
-    SheetSyncer.new.sync
+    SheetSync::Downloader.download
   end
 end
